Scientific Calculator Project
This scientific calculator project is built on top of my own String and
Tokenizer classes.  This is part of my Object-oriented design in C++
class assignment.

Can potentially add graphing feature.  I’m touching the surface of that
feature but still haven’t had time to finish it.
There are still bugs with the graphing part.  To be fixed later.

Using the Shuntingyard’s algorithm, this calculator translates natural
language to computer language, sends the data to the back-end
computing, commits the calculation and finally returns the result to
users.

Feature:
- Basic math calculation, such as +,-,*,/
- Reads mathematical variables
- Trigonometric calculation available.